+++ application-module/src/main/kotlin/com/reservation/timetable/exceptions/AllTheThingsAreAlreadyOccupiedException.kt
@@ -1,4 +1,4 @@
-package com.reservation.timetable.exception
+package com.reservation.timetable.exceptions
 
 import com.reservation.exceptions.ClientException
